refactor(mirror-registry): name application composition
