+++ internal/oci/boundary_test.go
@@ -45,8 +45,6 @@ func distributionBoundaryExempt(importPath string) bool {
 	short := strings.TrimPrefix(importPath, "github.com/quay/quay/")
 	return strings.HasPrefix(short, "internal/registry/") ||
 		short == "internal/registry" ||
-		strings.HasPrefix(short, "internal/quay/") ||
-		short == "internal/quay" ||
 		strings.HasPrefix(short, "internal/oci/storage/local/") ||
 		short == "internal/oci/storage/local"
 }
